Split out PathCleaner service
