+++ pathauto.services.yml
@@ -2,3 +2,8 @@ services:
   pathauto.manager:
     class: Drupal\pathauto\PathautoManager
     arguments: ['@config.factory', '@language_manager', '@cache.default', '@module_handler', '@token']
+  pathauto.alias_cleaner:
+    class: Drupal\pathauto\AliasCleaner
+    arguments: ['@config.factory', '@pathauto.alias_storage_helper']
+  pathauto.alias_storage_helper:
+    class: Drupal\pathauto\AliasStorageHelper
